+++ compiler/lambdas.cpp
@@ -50,13 +50,13 @@ static vector<TypePtr> typesOfValues(ObjectPtr obj) {
     switch (obj->objKind) {
     case MULTI_PVALUE : {
         MultiPValue *mpv = (MultiPValue *)obj.ptr();
-        for (unsigned i = 0; i < mpv->size(); ++i)
+        for (size_t i = 0; i < mpv->size(); ++i)
             types.push_back(mpv->values[i].type);
         break;
     }
     case MULTI_CVALUE : {
         MultiCValue *mcv = (MultiCValue *)obj.ptr();
-        for (unsigned i = 0; i < mcv->size(); ++i)
+        for (size_t i = 0; i < mcv->size(); ++i)
             types.push_back(mcv->values[i]->type);
         break;
     }
@@ -159,7 +159,7 @@ static void initializeLambdaWithFreeVars(LambdaPtr x, EnvPtr env,
     vector<RecordFieldPtr> fields;
 
     CallPtr converted = new Call(NULL, new ExprList());
-    for (unsigned i = 0; i < x->freeVars.size(); ++i) {
+    for (size_t i = 0; i < x->freeVars.size(); ++i) {
         IdentifierPtr ident = Identifier::get(x->freeVars[i]);
         NameRefPtr nameRef = new NameRef(ident);
 
@@ -184,7 +184,7 @@ static void initializeLambdaWithFreeVars(LambdaPtr x, EnvPtr env,
         case MULTI_CVALUE : {
             vector<TypePtr> types = typesOfValues(obj);
             vector<TypePtr> elementTypes;
-            for (unsigned j = 0; j < types.size(); ++j) {
+            for (size_t j = 0; j < types.size(); ++j) {
                 TypePtr t = types[j];
                 if (x->captureBy == REF_CAPTURE)
                     t = pointerType(t);
@@ -226,7 +226,7 @@ static void initializeLambdaWithFreeVars(LambdaPtr x, EnvPtr env,
     IdentifierPtr closureDataIdent = Identifier::get(closureDataName);
     FormalArgPtr closureDataArg = new FormalArg(closureDataIdent, typeExpr);
     code->formalArgs.push_back(closureDataArg.ptr());
-    for (unsigned i = 0; i < x->formalArgs.size(); ++i) {
+    for (size_t i = 0; i < x->formalArgs.size(); ++i) {
         code->formalArgs.push_back(x->formalArgs[i]);
     }
     code->hasVarArg = x->hasVarArg;
@@ -253,7 +253,7 @@ static void initializeLambdaWithoutFreeVars(LambdaPtr x, EnvPtr env,
 
     CodePtr code = new Code();
     code->location = x->location;
-    for (unsigned i = 0; i < x->formalArgs.size(); ++i) {
+    for (size_t i = 0; i < x->formalArgs.size(); ++i) {
         code->formalArgs.push_back(x->formalArgs[i]);
     }
     code->hasVarArg = x->hasVarArg;
@@ -295,7 +295,7 @@ void convertFreeVars(LambdaPtr x, EnvPtr env,
                      vector<string> &freeVars)
 {
     EnvPtr env2 = new Env(env);
-    for (unsigned i = 0; i < x->formalArgs.size(); ++i) {
+    for (size_t i = 0; i < x->formalArgs.size(); ++i) {
         FormalArgPtr arg = x->formalArgs[i];
         addLocal(env2, arg->name, arg->name.ptr());
     }
@@ -307,7 +307,7 @@ static EnvPtr convertFreeVarsFromBinding(BindingPtr binding, EnvPtr env, LambdaC
 {
     convertFreeVars(binding->values, env, ctx);
     EnvPtr env2 = new Env(env);
-    for (unsigned j = 0; j < binding->args.size(); ++j)
+    for (size_t j = 0; j < binding->args.size(); ++j)
         addLocal(env2, binding->args[j]->name, binding->args[j]->name.ptr());
     return env2;
 }
@@ -348,7 +348,7 @@ void convertFreeVars(StatementPtr x, EnvPtr env, LambdaContext &ctx)
 
     case BLOCK : {
         Block *y = (Block *)x.ptr();
-        for (unsigned i = 0; i < y->statements.size(); ++i) {
+        for (size_t i = 0; i < y->statements.size(); ++i) {
             StatementPtr z = y->statements[i];
             if (z->stmtKind == BINDING)
                 env = convertFreeVarsFromBinding((Binding *)z.ptr(), env, ctx);
@@ -407,7 +407,7 @@ void convertFreeVars(StatementPtr x, EnvPtr env, LambdaContext &ctx)
         Switch *y = (Switch *)x.ptr();
         env = convertFreeVarsFromStatementExpressionStatements(y->exprStatements, env, ctx);
         convertFreeVars(y->expr, env, ctx);
-        for (unsigned i = 0; i < y->caseBlocks.size(); ++i) {
+        for (size_t i = 0; i < y->caseBlocks.size(); ++i) {
             CaseBlockPtr z = y->caseBlocks[i];
             convertFreeVars(z->caseLabels, env, ctx);
             convertFreeVars(z->body, env, ctx);
@@ -446,7 +446,7 @@ void convertFreeVars(StatementPtr x, EnvPtr env, LambdaContext &ctx)
         For *y = (For *)x.ptr();
         convertFreeVars(y->expr, env, ctx);
         EnvPtr env2 = new Env(env);
-        for (unsigned j = 0; j < y->variables.size(); ++j)
+        for (size_t j = 0; j < y->variables.size(); ++j)
             addLocal(env2, y->variables[j], y->variables[j].ptr());
         convertFreeVars(y->body, env2, ctx);
         break;
@@ -459,7 +459,7 @@ void convertFreeVars(StatementPtr x, EnvPtr env, LambdaContext &ctx)
     case TRY : {
         Try *y = (Try *)x.ptr();
         convertFreeVars(y->tryBlock, env, ctx);
-        for (unsigned i = 0; i < y->catchBlocks.size(); ++i) {
+        for (size_t i = 0; i < y->catchBlocks.size(); ++i) {
             EnvPtr env2 = new Env(env);
             addLocal(env2,
                 y->catchBlocks[i]->exceptionVar,
@@ -569,7 +569,7 @@ void convertFreeVars(ExprPtr &x, EnvPtr env, LambdaContext &ctx)
             else if ((z->objKind == MULTI_PVALUE) || (z->objKind == MULTI_CVALUE)) {
                 vector<TypePtr> types = typesOfValues(z);
                 bool allStatic = true;
-                for (unsigned i = 0; i < types.size(); ++i) {
+                for (size_t i = 0; i < types.size(); ++i) {
                     if (!isStaticOrTupleOfStatics(types[i])) {
                         allStatic = false;
                         break;
@@ -577,7 +577,7 @@ void convertFreeVars(ExprPtr &x, EnvPtr env, LambdaContext &ctx)
                 }
                 if (allStatic) {
                     ExprListPtr args = new ExprList();
-                    for (unsigned i = 0; i < types.size(); ++i)
+                    for (size_t i = 0; i < types.size(); ++i)
                         args->add(new ObjectExpr(types[i].ptr()));
                     CallPtr call = new Call(operator_expr_typesToRValues(), args);
                     call->location = y->location;
@@ -682,7 +682,7 @@ void convertFreeVars(ExprPtr &x, EnvPtr env, LambdaContext &ctx)
     case LAMBDA : {
         Lambda *y = (Lambda *)x.ptr();
         EnvPtr env2 = new Env(env);
-        for (unsigned i = 0; i < y->formalArgs.size(); ++i)
+        for (size_t i = 0; i < y->formalArgs.size(); ++i)
             addLocal(env2, y->formalArgs[i]->name, y->formalArgs[i]->name.ptr());
         convertFreeVars(y->body, env2, ctx);
         break;
@@ -724,7 +724,7 @@ void convertFreeVars(ExprPtr &x, EnvPtr env, LambdaContext &ctx)
 
 void convertFreeVars(ExprListPtr x, EnvPtr env, LambdaContext &ctx)
 {
-    for (unsigned i = 0; i < x->size(); ++i)
+    for (size_t i = 0; i < x->size(); ++i)
         convertFreeVars(x->exprs[i], env, ctx);
 }
